refactor: separate out helper functions into own files
